+++ server-spi/src/main/java/org/keycloak/models/RealmModel.java
@@ -255,6 +255,11 @@ public interface RealmModel extends RoleContainerModel {
 
     void setClaimsSupported(List<String> claimsSupported);
 
+    String getDefaultAudValueForAccessToken();
+
+    void setDefaultAudValueForAccessToken(String defaultAudValueForAccessToken);
+
+
     OAuth2DeviceConfig getOAuth2DeviceConfig();
 
     CibaConfig getCibaPolicy();