+++ services/src/main/java/org/keycloak/protocol/oidc/grants/device/DeviceGrantType.java
@@ -60,7 +60,8 @@ import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
 
 import java.net.URI;
-import java.util.stream.Stream;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author <a href="mailto:h2-wada@nri.co.jp">Hiroyuki Wada</a>
@@ -213,6 +214,9 @@ public class DeviceGrantType {
                 "The authorization request is still pending", Response.Status.BAD_REQUEST);
         }
 
+        List<String> resourceList = formParams.get(OAuth2Constants.RESOURCE);
+        TokenEndpoint.checkResourceList(resourceList, event, cors);
+
         // https://tools.ietf.org/html/rfc7636#section-4.6
         String codeVerifier = formParams.getFirst(OAuth2Constants.CODE_VERIFIER);
         String codeChallenge = deviceCodeModel.getCodeChallenge();
@@ -298,6 +302,6 @@ public class DeviceGrantType {
         // Set nonce as an attribute in the ClientSessionContext. Will be used for the token generation
         clientSessionCtx.setAttribute(OIDCLoginProtocol.NONCE_PARAM, deviceCodeModel.getNonce());
 
-        return tokenEndpoint.createTokenResponse(user, userSession, clientSessionCtx, scopeParam, false);
+        return tokenEndpoint.createTokenResponse(user, userSession, clientSessionCtx, scopeParam, false, resourceList);
     }
 }