+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/util/OAuthClient.java
@@ -185,6 +185,8 @@ public class OAuthClient {
 
     private String requestUri;
 
+    private String resource;
+
     private String claims;
 
     private Map<String, String> requestHeaders;
@@ -500,6 +502,9 @@ public class OAuthClient {
         // https://tools.ietf.org/html/rfc7636#section-4.5
         if (codeVerifier != null) {
             parameters.add(new BasicNameValuePair(OAuth2Constants.CODE_VERIFIER, codeVerifier));
+        }
+        if (resource != null) {
+            parameters.add(new BasicNameValuePair(OAuth2Constants.RESOURCE, resource));
         }
 
         UrlEncodedFormEntity formEntity = new UrlEncodedFormEntity(parameters, Charsets.UTF_8);
@@ -998,6 +1003,9 @@ public class OAuthClient {
         if (clientSessionHost != null) {
             parameters.add(new BasicNameValuePair(AdapterConstants.CLIENT_SESSION_HOST, clientSessionHost));
         }
+        if (resource != null) {
+            parameters.add(new BasicNameValuePair(OAuth2Constants.RESOURCE, resource));
+        }
 
         UrlEncodedFormEntity formEntity;
         try {
@@ -1041,6 +1049,9 @@ public class OAuthClient {
             }
             if (codeChallengeMethod != null) {
                 parameters.add(new BasicNameValuePair(OAuth2Constants.CODE_CHALLENGE_METHOD, codeChallengeMethod));
+            }
+            if (resource != null) {
+                parameters.add(new BasicNameValuePair(OAuth2Constants.RESOURCE, resource));
             }
 
             UrlEncodedFormEntity formEntity;
@@ -1478,6 +1489,9 @@ public class OAuthClient {
         if (codeChallengeMethod != null) {
             b.queryParam(OAuth2Constants.CODE_CHALLENGE_METHOD, codeChallengeMethod);
         }
+//        if (resource != null) {
+//            b.queryParam(OAuth2Constants.RESOURCE, resource);
+//        }
         if (customParameters != null) {
             customParameters.keySet().stream().forEach(i -> b.queryParam(i, customParameters.get(i)));
         }
@@ -1614,6 +1628,11 @@ public class OAuthClient {
 
     public OAuthClient redirectUri(String redirectUri) {
         this.redirectUri = redirectUri;
+        return this;
+    }
+
+    public OAuthClient resource(String resource) {
+        this.resource = resource;
         return this;
     }